+++ src/test/windows.spec.ts
@@ -1,55 +1,55 @@
 import chai from "chai";
-import { parseWindows, ScriptUrl } from "../lib";
+import isWindows from "is-windows";
+import { ParsedScriptUrl, parseSys, parseWindows } from "../lib";
 
 interface TestItem {
   url: string;
-  expected: ScriptUrl;
+  expected: ParsedScriptUrl;
 }
 
 const testItems: ReadonlyArray<TestItem> = [
   {
     url: "",
-    expected: {isRegularFile: false, scriptUrl: ""},
+    expected: {url: "", isFileUrl: false},
   },
   {
     url: "url.js",
-    expected: {isRegularFile: false, scriptUrl: "url.js"},
+    expected: {url: "url.js", isFileUrl: false},
   },
   {
     url: "internal/url.js",
-    expected: {isRegularFile: false, scriptUrl: "internal/url.js"},
+    expected: {url: "internal/url.js", isFileUrl: false},
   },
   {
     url: "internal/util/inspect.js",
-    expected: {isRegularFile: false, scriptUrl: "internal/util/inspect.js"},
+    expected: {url: "internal/util/inspect.js", isFileUrl: false},
   },
   {
     url: "C:\\foo.js",
-    expected: {
-      isRegularFile: true,
-      scriptUrl: "C:\\foo.js",
-      moduleType: "cjs",
-      url: "file:///C:/foo.js",
-      path: "C:\\foo.js",
-    },
+    expected: {url: "C:\\foo.js", isFileUrl: false},
   },
   {
     url: "file:///C:/foo.mjs",
-    expected: {
-      isRegularFile: true,
-      scriptUrl: "file:///C:/foo.mjs",
-      moduleType: "esm",
-      url: "file:///C:/foo.mjs",
-      path: "C:\\foo.mjs",
-    },
+    expected: {url: "file:///C:/foo.mjs", isFileUrl: true, path: "C:\\foo.mjs"},
   },
 ];
 
 describe("parseWindows", () => {
   for (const {url, expected} of testItems) {
     it(JSON.stringify(url), () => {
-      const actual: ScriptUrl = parseWindows(url);
+      const actual: ParsedScriptUrl = parseWindows(url);
       chai.assert.deepEqual(actual, expected);
     });
   }
 });
+
+function parseSysSuite() {
+  for (const {url, expected} of testItems) {
+    it(JSON.stringify(url), () => {
+      const actual: ParsedScriptUrl = parseSys(url);
+      chai.assert.deepEqual(actual, expected);
+    });
+  }
+}
+
+(isWindows() ? describe : describe.skip)("parseSys (windows)", parseSysSuite);