+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+
+- **[Breaking change]** Update lib for Node 11. The new result only has
+  `url` (input), `isFileUrl` (boolean indicating a regular `file://` URL) and
+  an eventual `path`.
+
 # 0.1.2 (2018-10-20)
 
 - **[Fix]** Fix OS detection